pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-code"
version = "0.1.0"
description = "Open Source CLI Coding Agent"
authors = [
    { name = "Open Source CLI Coding Agent Contributors" }
]
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "typer",
    "httpx",
    "pydantic",
    "gitpython",
    "toml",
    "pyyaml"
]

[project.scripts]
ai-code = "src.cli.main:app"
